+++ R/options.R
@@ -1,0 +1,20 @@
+#' Package options
+#'
+#' Options that adjust the behaviour of the palettes package.
+#'
+#' These options can be set via [options()] and queried via [getOption()].
+#'
+#' @section Options for the palettes package:
+#'
+#' \describe{
+#'
+#' \item{\code{palettes.symbol}:}{
+#'
+#'   Character string setting the symbol used for colour previews. See
+#'   `cli::list_symbols()` for a list of symbol choices. Defaults to `"bullet"`.
+#'
+#' }
+#'
+#' }
+#' @name palettes-options
+NULL