+++ R/ggplot2-scales.R
@@ -1,8 +1,11 @@
 #' Colour scales from colour vectors and colour palettes
 #'
+#' Create discrete, continuous, and binned colour scales from colour vectors and
+#' colour palettes.
+#'
 #' @inheritParams pal_ramp
-#' @param ... Other arguments passed on to `ggplot2::discrete_scale()`,
-#'   `ggplot2::continuous_scale()`, or `ggplot2::binned_scale()` to control name,
+#' @param ... Other arguments passed on to [ggplot2::discrete_scale()],
+#'   [ggplot2::continuous_scale()], or [ggplot2::binned_scale()] to control name,
 #'   limits, breaks, labels and so forth.
 #' @return A scale function that controls the mapping between data and colour or
 #'   fill aesthetics in a [ggplot2][ggplot2::ggplot2-package] plot.