+++ NEWS.md
@@ -2,6 +2,8 @@
 
 - Added the `palettes.symbol` option to set the symbol used for colour previews. See `?"palettes-options"` for details.
 
+- Improved error message for invalid colours to support singular and plural grammar (@olivroy, #32)
+
 # palettes 0.1.1
 
 This is a small patch release that fixes a typo in the "Creating a Colour Package" vignette, replacing the non-existent `pal_brewer()` function with the intended `pal_ramp()` function.